Create base test class and update existing test
