+++ firefox-ios/firefox-ios-tests/Tests/ClientTests/BrowserViewController/Constraints/BrowserViewControllerConstraintTests.swift
@@ -8,26 +8,7 @@ import XCTest
 @testable import Client
 
 @MainActor
-final class BrowserViewControllerConstraintTests: XCTestCase {
-    var profile: MockProfile!
-    var tabManager: MockTabManager!
-
-    override func setUp() async throws {
-        try await super.setUp()
-        tabManager = MockTabManager()
-        DependencyHelperMock().bootstrapDependencies(injectedTabManager: tabManager)
-        profile = MockProfile()
-        LegacyFeatureFlagsManager.shared.initializeDeveloperFeatures(with: profile)
-    }
-
-    override func tearDown() async throws {
-        profile.shutdown()
-        profile = nil
-        tabManager = nil
-        DependencyHelperMock().reset()
-        try await super.tearDown()
-    }
-
+final class BrowserViewControllerConstraintTests: BrowserViewControllerConstraintTestsBase {
     // MARK: - Container View Existence Tests
 
     func test_containerViews_existInHierarchy() {
@@ -288,29 +269,6 @@ final class BrowserViewControllerConstraintTests: XCTestCase {
     }
 
     // MARK: - Helper Methods
-
-    private func createSubject(isFeatureFlagEnabled: Bool = false) -> BrowserViewController {
-        // Setup feature flag to disabled by default and override only in the test that need it
-        setupNimbusSnapKitRemovalTesting(isEnabled: isFeatureFlagEnabled)
-        let subject = BrowserViewController(profile: profile,
-                                            tabManager: tabManager)
-        trackForMemoryLeaks(subject)
-
-        // Trigger view loading and constraint setup
-        // SnapKit constraints are created in updateViewConstraints(), so we need to explicitly trigger it
-        subject.loadViewIfNeeded()
-        subject.view.setNeedsUpdateConstraints()
-        subject.view.updateConstraintsIfNeeded()
-        subject.view.layoutIfNeeded()
-
-        return subject
-    }
-
-    private func setupNimbusSnapKitRemovalTesting(isEnabled: Bool) {
-        FxNimbus.shared.features.snapkitRemovalRefactor.with { _, _ in
-            return SnapkitRemovalRefactor(enabled: isEnabled)
-        }
-    }
 
     /// Check if a view has a constraint with the given attribute related to another view
     private func hasConstraint(for view: UIView,
